refactor: Relocate getBlockForTimestamp()
This function uses Redis in the back-end to cache block timestamps; it's
not inherently a part of Redis utilities. Also, export getBlockFinder()
so that it can be used directly.
